pyproject.toml
[project]
name = "search-first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langgraph-cli[inmem]>=0.4.2",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "yfinance>=0.2.66",
]
